url handling | app list improvement
