--- src/components/app/list/KubernetessListContainer.tsx
@@ -1,11 +1,0 @@
-import React, { Component } from 'react'
-
-export default class k8sListContainer extends Component {
-    render() {
-        return (
-            <div>
-                
-            </div>
-        )
-    }
-}
