+++ src/components/app/list/ListContainer.tsx
@@ -2,25 +2,22 @@ import React, { Component } from 'react'
 import { Switch, Route, Redirect, NavLink } from 'react-router-dom';
 import AppListContainer from './AppListContainer';
 import ExternalListContainer from './ExternalListContainer';
-import { ListContainerProps, ListContainerState } from './types';
-import { AppListViewType } from '../config';
+import { ListContainerProps } from './types';
 import { URLS } from '../../../config';
 import { AddNewApp } from '../create/CreateApp';
+import { ReactComponent as Add } from '../../../assets/icons/ic-add.svg';
+import './list.css';
 
 const APP_LIST_PARAM = {
     createApp: 'create-app',
 }
 
-export default class ListContainer extends React.Component<ListContainerProps, ListContainerState> {
+export default class ListContainer extends Component<ListContainerProps, {}> {
     constructor(props) {
-        super(props)
+        super(props);
 
-        this.state = {
-            collapsed: false,
-            code: 0,
-            view: AppListViewType.LOADING,
-        }
         this.handleKeyPress = this.handleKeyPress.bind(this);
+        this.openCreateModal = this.openCreateModal.bind(this);
     }
 
     componentDidMount() {
@@ -28,21 +25,19 @@ export default class ListContainer extends React.Component<ListContainerProps, L
     }
 
     handleKeyPress(event) {
-        var activeElement = document.activeElement
-        var inputs = ['input', 'select', 'button', 'textarea'];
-
+        let activeElement = document.activeElement
+        let inputs = ['input', 'select', 'button', 'textarea'];
         if (activeElement && inputs.indexOf(activeElement.tagName.toLowerCase()) !== -1) {
             return;
         }
 
-        if (event.key === 'E') {
-            let url = `${URLS.APP}/external-apps`
+        if (event.key === 'E' || event.key === 'e') {
+            let url = `${URLS.EXTERNAL_APP}`;
             this.props.history.push(`${url}`);
         }
-        else if (event.key === 'A') {
-            let url = `${URLS.APP}`
+        else if (event.key === 'A' || event.key === 'a') {
+            let url = `${URLS.APP}`;
             this.props.history.push(`${url}`);
-            { console.log(document.hasFocus(), document.activeElement) }
         }
     }
 
@@ -50,49 +45,43 @@ export default class ListContainer extends React.Component<ListContainerProps, L
         document.removeEventListener("keydown", this.handleKeyPress);
     }
 
-    openCreateModal = (event: React.MouseEvent): void => {
-        let url = `${URLS.APP}/${APP_LIST_PARAM.createApp}${this.props.location.search}`
+    openCreateModal(event: React.MouseEvent): void {
+        let url = `${URLS.APP}/${APP_LIST_PARAM.createApp}${this.props.location.search}`;
         this.props.history.push(`${url}`);
     }
 
     renderListHeader() {
-        const path = this.props.match.path;
-        let url = `${window.location.origin}/external-apps`;
-        return <>
-            <ul role="tablist" className="tab-list border-btm bcn-0 pl-20">
-                <li className="tab-list__tab ellipsis-right">
-                    <NavLink activeClassName="active" to={`${path}`} className="tab-list__tab-link">Devtron Apps [A]</NavLink>
-                </li>
-                <li className="tab-list__tab">
-                    <NavLink activeClassName="active" to={`${URLS.APP}/external-apps`} className="tab-list__tab-link">External Apps [E]</NavLink>
-                </li>
-                <li className="tab-list__tab--right mt-9 mb-9 mr-20"><button type="button" className="cta h-30"
-                    onClick={this.openCreateModal}>
-                    <span className="round-button__icon"><i className="fa fa-plus" aria-hidden="true"></i></span>
-                     New app
-                </button></li>
-            </ul>
-        </>
+        return <ul role="tablist" className="tab-list border-btm bcn-0 pl-20">
+            <li className="tab-list__tab ellipsis-right">
+                <NavLink activeClassName="active" to={`${URLS.APP}`} className="tab-list__tab-link">Devtron Apps [A]</NavLink>
+            </li>
+            <li className="tab-list__tab">
+                <NavLink activeClassName="active" to={`${URLS.EXTERNAL_APP}`} className="tab-list__tab-link">External Apps [E]</NavLink>
+            </li>
+            <li className="tab-list__tab--right mt-9 mb-9 mr-20">
+                <button type="button" className="cta h-30" onClick={this.openCreateModal}>
+                    <Add className="" />New app
+                </button>
+            </li>
+        </ul>
     }
 
     renderRouter() {
-        const path = this.props.match.path
         return <Switch>
-            <Route exact path={`${path}`} component={AppListContainer} />
-            <Route path={`${URLS.APP}/external-apps`} component={ExternalListContainer} />
+            <Route path={`${URLS.APP}`} component={AppListContainer} />
+            <Route path={`${URLS.EXTERNAL_APP}`} component={ExternalListContainer} />
             <Route path={`${URLS.APP}/${APP_LIST_PARAM.createApp}`}
-                render={(props) => <AddNewApp close={this.props.closeModal} match={props.match} location={props.location} history={props.history} />}
+                render={(props) => <AddNewApp close={this.props.closeModal}
+                    match={props.match} location={props.location} history={props.history} />}
             />
-            <Redirect exact to={`${path}`} />
+            <Redirect to={`${URLS.APP}`} />
         </Switch>
     }
 
     render() {
-        return (
-            <div>
-                {this.renderListHeader()}
-                {this.renderRouter()}
-            </div>
-        )
+        return <div>
+            {this.renderListHeader()}
+            {this.renderRouter()}
+        </div>
     }
 }
